Extract createProjectFromTemplate/createProjectWithoutTemplate to own files
